+++ build_helpers/chroot/build_static_local.sh
@@ -8,9 +8,9 @@
 #  * Install binfmt support and qemu to enable running non-native executables:
 #    $ sudo apt-get install qemu binfmt-support qemu-user-static
 #  * Register multiarch support:
-#    $ docker run --rm --privileged multiarch/qemu-user-static --reset -p yes 
+#    $ docker run --rm --privileged multiarch/qemu-user-static --reset -p yes
 #  * Try running an arm64 executable:
-#    $ docker run --rm -t arm64v8/ubuntu uname -m 
+#    $ docker run --rm -t arm64v8/ubuntu uname -m
 
 
 CHROOT_PATH=${CHROOT_PATH:="/var/tmp/elbencho_chroot_$(whoami)"}
@@ -85,14 +85,14 @@ sudo "$ALPINE_SCRIPT_PATH" -b "$CHROOT_VERSION" -a "$BUILD_ARCH" \
   -d "$(realpath "$CHROOT_PATH")" -i $(pwd) -p \
   "bash boost-dev build-base gcc g++ git libaio-dev make numactl-dev \
         cmake curl-dev curl-static openssl-libs-static ncurses-static \
-        boost-static ncurses zlib-static libretls-static nghttp2-static \
+        boost-static ncurses zlib-static libretls-static nghttp2-static nghttp3-static \
         brotli-static ncurses-dev sudo tar libidn2-static libunistring-static \
         libpsl-static c-ares-dev zstd-static"
 
 if [ $? -ne 0 ]; then
   echo "ERROR: Preparation of chroot failed."
   cleanup_chroot_and_exit
-  exit 1 
+  exit 1
 fi
 
 echo
@@ -105,7 +105,7 @@ sudo "$CHROOT_PATH/enter-chroot" -u $(whoami) make -C "$(pwd)" -j $NUM_JOBS \
 if [ $? -ne 0 ]; then
   echo "ERROR: Build failed."
   cleanup_chroot_and_exit
-  exit 1 
+  exit 1
 fi
 
 bin/elbencho --version && \
@@ -116,7 +116,7 @@ bin/elbencho --version && \
 if [ $? -ne 0 ]; then
   echo "ERROR: Executable packaging failed."
   cleanup_chroot_and_exit
-  exit 1 
+  exit 1
 fi
 
 echo