+++ source/workers/Worker.cpp
@@ -12,6 +12,10 @@ void Worker::threadStart(Worker* worker)
 {
 	SignalTk::registerFaultSignalHandlers(*(worker->progArgs) );
 
+    // set thread name (max 15 chars plus '\0')
+    std::string threadName = "elb-wrk-" + std::to_string(worker->workerRank);
+    pthread_setname_np(pthread_self(), threadName.c_str() );
+
 	worker->run();
 	worker->cleanup();
 }