+++ CHANGELOG.md
@@ -7,14 +7,18 @@
 * Added option for Excel (`.xlsx`) file generation to the elbencho chart tool. (`elbencho-chart --excel PATH`)
 * Added tool to clean up unfinished S3 multi-part uploads. (See `tools/s3-cleanup-mpu.py`.)
 * Added support for multiple S3 credentials. (See `--s3credfile` & `--s3credlist`.)
+* Added option to use a single shared S3 client instance instead of separate instances for each worker thread. (See `--s3single`.)
 
 ### General Changes
 * Added config values to `--jsonfile` output.
 * Removed dependency on lib boost_system for compatibility with more recent boost versions.
 * Added `nvidia-open` package to CUDA container to have `nvidia-smi` tool available inside container.
 
+### Fixes
+* Disabled S3 continue request handler (which allowed for faster termination in case of errors or user interruption) when built with S3_AWSCRT=1 as a workaround for this issue: https://github.com/aws/aws-sdk-cpp/issues/3639
+
 ### Contributors
-* Thanks to John Fragalla, Tamar Baran, Michael Shustin & Github user panghubaobao777 for contributions, helpful comments and suggestions.
+* Thanks to John Fragalla, Kiran Modukuri, Tamar Baran, Michael Shustin & GitHub user panghubaobao777 for contributions, helpful comments and suggestions.
 
 ## v3.0.35 (Sep 1, 2025)